requirements.txt
annotated-types==0.6.0
appdirs==1.4.4
apturl==0.5.2
attrs==21.2.0
autopage==0.4.0
Babel==2.8.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cliff==3.10.0
cmd2==0.0.0
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debtcollector==2.3.0
decorator==4.4.2
defer==1.0.6
deprecation==2.0.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
dogpile.cache==1.1.4
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.1
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iso8601==1.0.2
jeepney==0.7.1
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.0
keyring==23.5.0
keystoneauth1==4.4.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
munch==2.5.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openstacksdk==0.61.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.5.0
oslo.config==8.8.0
oslo.context==4.1.0
oslo.i18n==5.1.0
oslo.log==4.6.1
oslo.serialization==4.2.0
oslo.utils==4.12.2
packaging==21.3
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
prettytable==2.5.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-cinderclient==8.3.0
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-keystoneclient==4.4.0
python-novaclient==17.6.0
python-openstackclient==5.8.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requestsexceptions==1.4.0
rfc3986==1.5.0
roman==3.3
rsa==4.9
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
speedtest-cli==2.1.3
ssh-import-id==5.11
stevedore==3.5.0
systemd-python==234
terminator==2.1.1
tqdm==4.66.2
typing_extensions==4.11.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
xdg==5
xkit==0.0.0
zipp==1.0.0
